chat_assistant/requirements.txt
pyyaml
appdirs==1.4.3
Django==2.2
djangorestframework==3.12.1
django-rest-swagger==2.2.0
packaging==20.0
pyparsing==2.2.0
pytz==2019.1
six==1.12.0
spacy
psycopg2==2.7.3.2
torch
uWSGI==2.0.18
django-jsonfield==1.4.1
coloredlogs==14.3
gast==0.3.3
httpcore==0.11.1
httplib2==0.18.1
multidict==5.0
numpy==1.19.5
rasa==2.2.4
rasa-sdk==2.2.0
requests==2.25.1
tqdm==4.31
celery==4.3.1
redis==3.5.3
django-cors-headers==3.7.0
drf-writable-nested==0.6.2
yarl==1.0